refactor documentation
